First draft for returning invoice to basket
This commit includes the basic functionality to revert invoice back to basket.
The functionality needs to be re-checked before I can confirm that it actually
works, but so far it seems to work. Additionally there's still some data formatting
and display which needs some refining, but so far things seem to work.

refs 49