More fixes to time tracking
Previous issue with this is really, really strange. If I make
an somewhat simple statement:
	if($_[2]=~/^[0-9]+$/ && $_[2]<0) {
the whole program freezes. But this seems to work:
	if(!($_[2]=~/\d\d\d\d-\d\d-\d\d/) && $_[2]<0) {

This can be seen in action at commit
885ea0c30949db1d7812114fda0299bd86566513

Go to tracking and see how everything freezes, I tracked it down
to statements above, if anyone has any clue, please let me know.